+++ src/main/java/jp/brainjuice/pokego/business/service/manage/masterFileAnalyzer/dto/json/QuickCombatMove.java
@@ -11,7 +11,7 @@ public class QuickCombatMove implements Serializable, Cloneable {
 
 	private String uniqueId;
 	private String type;
-	private float power;
+	private double power;
 	private String vfxName;
 	private int durationTurns;
 	private int energyDelta;
